+++ stream247/__init__.py
@@ -1,0 +1,3 @@
+from .runtime import main
+
+__all__ = ["main"]
