+++ stream247/__main__.py
@@ -1,0 +1,7 @@
+try:
+    from .runtime import main
+except ImportError:
+    from stream247.runtime import main
+
+if __name__ == "__main__":
+    main()
