+++ stream247/shared.py
@@ -1,0 +1,8 @@
+"""Backward-compatible re-export module for legacy imports."""
+
+from .constants import *
+from .config import *
+from .models import *
+from .qtcompat import QtCore
+from .updates import *
+from .utils import *
